db connects only once
