+++ .env.sample
@@ -1,3 +1,3 @@
-CLUSTER_NAME=something
-CLUSTER_PASSWORD=asdasdqweqwe
-DB_URI=dbname.somedata.mongodb.net/collectionname
+DB_ADMIN_NAME=something
+DB_ADMIN_PASSWORD=asdasdqweqwe
+DB_URI=thedburi
