+++ README.md
@@ -1,2 +1,8 @@
 # to-do-server
-Backend for the todo client
+Backend for the todo client, hosted at https://to-do-server1.herokuapp.com/
+
+# run server locally
+`npm start` to run server with `node app.js` or `nodemon app.js` if you want to the server reload if you make changes on the files
+
+# about the app
+A node.js REST API, written in express.js framework, that using mongodb NOSQL database.
