+++ app.js
@@ -1,7 +1,8 @@
 const express = require('express');
 const app = express();
-const db = require('./db');
 const todo = require('./controllers/todo');
+const MongoClient = require('mongodb').MongoClient;
+let dbClient;
 
 if (process.env.NODE_ENV !== 'production') {
   require('dotenv').config();
@@ -10,10 +11,20 @@ if (process.env.NODE_ENV !== 'production') {
 app.use(require('cors')());
 app.use(express.json());
 /* app.use(express.urlencoded({ extended: true })) */
-app.use(db.createConnection);
+
+app.use((req, res, next) => {
+  req.db = dbClient.db('todo');
+  next();
+});
 
 app.use('/todo', todo);
 
-app.use(db.closeConnection);
-
-app.listen(process.env.PORT || 8080);
+MongoClient.connect(
+  `mongodb+srv://${process.env.DB_ADMIN_NAME}:${process.env.DB_ADMIN_PASSWORD}@${process.env.DB_URI}?retryWrites=true&w=majority`,
+  { useUnifiedTopology: true },
+  function (err, client) {
+    if (err) throw err;
+    dbClient = client;
+    app.listen(process.env.PORT || 8080);
+  }
+);
